requirements.txt
python-telegram-bot
requests
python-dotenv
redis[hiredis]